requirements.txt
fastapi==0.116.1
geopy==2.4.1
pandas==2.3.2
passlib==1.7.4
python_jose==3.5.0
Requests==2.32.5
scikit_learn==1.7.1
SQLAlchemy==2.0.25
google-generativeai==0.8.5
python-dotenv==1.1.1
uvicorn==0.35.0
